requirements.txt
Flask
flask-sqlalchemy
flask-migrate
python-dotenv
flask-dance
mysql-connector-python
Flask-login